requirements.txt
Django==1.11.16
django-autoslug==1.9.3
Pillow==5.3.0
django-guardian==1.4.9
django-userena-ce==4.1.0
easy-thumbnails==2.5
django-extensions==2.1.3
django-parler==1.9.2
